record-api/requirements.txt
Flask==3.0.3
redis==5.0.4
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv